Functions exported from FourierFlows:

Functions exported from FourierFlows:

Base.resize! — Method.

resize!(diag, newnum)

Resize the Diagnostic data and time arrays to length newnum.

source

FourierFlows.increment! — Method.

increment!(diags)

Increment the array of Diagnostics diags.

source

FourierFlows.increment! — Method.

increment!(diag)

Increment the Diagnostic diag.

source

FourierFlows.savediagnostic — Method.

savediagnostic(diag, diagname)

Save diagnostics to file, labeled by the string diagname.

source

FourierFlows.saveoutput — Method.

saveoutput(out)

Save current output fields for file in out.filename.

source

FourierFlows.saveproblem — Method.

saveproblem(prob, filename)

Save certain aspects of a problem timestepper, grid, and params. Functions that are fields in params are not saved.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, diags, nsteps)

Step forward the problem prob for nsteps while calculating the diagnostics in diags.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, nsteps)

Step forward the problem 'prob' for 'nsteps'.

source

FourierFlows.stepforward! — Method.

stepforward!(prob)

Step forward the Problem prob for one timestep.

source

FourierFlows.update! — Method.

update!(diag)

Update diag with its current value.

source

Private types in module FourierFlows:

FourierFlows.ZeroDGrid — Type.

Doc.

source