Functions

Functions

Functions exported from FourierFlows:

Base.resize! — Method.

resize!(diag, newnum)

Resize the Diagnostic data and time arrays to length newnum.

source

FourierFlows.increment! — Method.

increment!(diag)
increment!(diags)

Increment the Diagnostic diag, or an array of Diagnostics diags.

source

FourierFlows.savediagnostic — Method.

savediagnostic(diag, diagname)

Save diagnostics to file, labeled by the string diagname.

source

FourierFlows.saveoutput — Method.

saveoutput(out)

Save current output fields for file in out.filename.

source

FourierFlows.saveproblem — Method.

saveproblem(prob, filename)

Save certain aspects of a problem timestepper, grid, and params. Functions that are fields in params are not saved.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, diags, nsteps)

Step forward prob for nsteps, incrementing diagnostics in the array diags along the way.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, nsteps)

Step forward prob for nsteps.

source

FourierFlows.stepforward! — Method.

stepforward!(prob)

Step forward the Problem prob for one timestep.

source

FourierFlows.update! — Method.

update!(diag)

Update diag with its current value.

source

Functions exported from KuramotoSivashinsky:

FourierFlows.KuramotoSivashinsky.InitialValueProblem — Method.

InitialValueProblem(; parameters...)

Construct an initial-value Kuramoto-Sivashinky problem that solves the equation

∂t u + ∂ₓ⁴u + ∂ₓ²u + u ∂ₓu = 0.

source

FourierFlows.KuramotoSivashinsky.set_u! — Method.

set_u!(prob, u)
set_u!(s, v, g, u)

Set the solution prob.state.sol as the transform of u and update variables.

source

FourierFlows.KuramotoSivashinsky.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

source

Functions exported from TwoDTurb:

Functions exported from BarotropicQG:

Functions exported from TracerAdvDiff:

Functions exported from VerticallyFourierBoussinesq:

Functions exported from VerticallyCosineBoussinesq:

FourierFlows.VerticallyCosineBoussinesq.mode0apv — Method.

mode0apv(prob)

Returns the barotropic available potential vorticity.

source

FourierFlows.VerticallyCosineBoussinesq.mode0apv — Method.

mode0apv(uh, vh, ph, Zh, m, N, g)

Returns the barotropic available potential vorticity.

source

FourierFlows.VerticallyCosineBoussinesq.mode0dissipation — Method.

mode0dissipation(prob)

Returns the domain-averaged barotropic dissipation rate. nnu0 must be >= 1.

source

FourierFlows.VerticallyCosineBoussinesq.mode0drag — Method.

mode0drag(prob)

Returns the extraction of domain-averaged barotropic energy by drag μ.

source

FourierFlows.VerticallyCosineBoussinesq.mode0energy — Method.

mode0energy(prob)

Returns the domain-averaged energy in the zeroth mode.

source

FourierFlows.VerticallyCosineBoussinesq.mode0enstrophy — Method.

mode0enstrophy(prob)

Returns the domain-averaged enstrophy in the Fourier-transformed vorticity solution s.sol.

source

FourierFlows.VerticallyCosineBoussinesq.mode1dissipation — Method.

mode1dissipation(prob)

Returns the domain-averaged kinetic energy dissipation of the first mode by horizontal viscosity.

source

FourierFlows.VerticallyCosineBoussinesq.mode1drag — Method.

mode1drag(prob)

Returns the domain-averaged kinetic energy dissipation of the first mode by horizontal viscosity.

source

FourierFlows.VerticallyCosineBoussinesq.mode1energy — Method.

mode1energy(prob)

Returns the domain-averaged total energy in the first mode.

source

FourierFlows.VerticallyCosineBoussinesq.mode1ke — Method.

mode1ke(prob)

Returns the domain-averaged kinetic energy in the first mode.

source

FourierFlows.VerticallyCosineBoussinesq.mode1pe — Method.

mode1pe(prob)

Returns the domain-averaged potential energy in the first mode.

source

FourierFlows.VerticallyCosineBoussinesq.set_C! — Method.

set_C!(prob, C)

Set zeroth mode tracer concentration and update vars.

source

FourierFlows.VerticallyCosineBoussinesq.set_Z! — Method.

set_Z!(prob, Z)

Set zeroth mode vorticity and update vars.

source

FourierFlows.VerticallyCosineBoussinesq.set_planewave! — Function.

set_planewave!(prob, u₀, κ, θ=0)

Set a plane wave solution with initial speed u₀, non-dimensional wave number κ, and angle θ with the horizontal. The non-dimensional wavenumber vector is (k, l) = (κ cos θ, κ sin θ), is normalized by 2π/Lx, and is rounded to the nearest integer.

source

FourierFlows.VerticallyCosineBoussinesq.set_uvp! — Method.

set_uvp!(prob)

Set first mode u, v, and p and update vars.

source

FourierFlows.VerticallyCosineBoussinesq.totalenergy — Method.

totalenergy(prob)

Returns the total energy projected onto the zeroth mode.

source

FourierFlows.VerticallyCosineBoussinesq.updatevars! — Method.

updatevars!(prob)

Update variables to correspond to the solution in s.sol or prob.state.sol.

source