Private types

Private types

Private types in module FourierFlows:

FourierFlows.ZeroDGrid — Type.

ZeroDGrid()

Constructs a placeholder grid object for "0D" problems (in other words, systems of ODEs).

source

Private types in module KuramotoSivashinsky:

FourierFlows.KuramotoSivashinsky.Vars — Method.

Returns the Vars object for Kuramoto-Sivashinsky.

source

Private types in module TwoDTurb:

FourierFlows.TwoDTurb.ForcedParams — Type.

ForcedParams(nu, nnu, mu, nmu, calcF!)

Returns the params for forced two-dimensional turbulence.

source

FourierFlows.TwoDTurb.ForcedVars — Method.

ForcedVars(g)

Returns the vars for forced two-dimensional turbulence with grid g.

source

FourierFlows.TwoDTurb.Params — Type.

Params(nu, nnu, mu, nmu)

Returns the params for unforced two-dimensional turbulence.

source

FourierFlows.TwoDTurb.Vars — Method.

Vars(g)

Returns the vars for unforced two-dimensional turbulence with grid g.

source

Private types in module BarotropicQG:

FourierFlows.BarotropicQG.ForcedParams — Type.

ForcedParams(g::TwoDGrid, f0, beta, FU, eta, mu, nu, nnu)

Returns the params for an forced two-dimensional barotropic QG problem.

source

FourierFlows.BarotropicQG.ForcedParams — Method.

ForcedParams(g::TwoDGrid, f0, beta, eta::Function, mu, nu, nnu, calcFU, calcFq)

Constructor for Params that accepts a generating function for the topographic PV.

source

FourierFlows.BarotropicQG.ForcedVars — Type.

ForcedVars(g)

Returns the vars for forced two-dimensional barotropic QG problem with grid g.

source

FourierFlows.BarotropicQG.Params — Type.

Params(g::TwoDGrid, f0, beta, FU, eta, mu, nu, nnu)

Returns the params for an unforced two-dimensional barotropic QG problem.

source

FourierFlows.BarotropicQG.Params — Method.

Params(g::TwoDGrid, f0, beta, eta::Function, mu, nu, nnu)

Constructor for Params that accepts a generating function for the topographic PV.

source

FourierFlows.BarotropicQG.Vars — Type.

Vars(g)

Returns the vars for unforced two-dimensional barotropic QG problem with grid g.

source

Private types in module TracerAdvDiff:

FourierFlows.TracerAdvDiff.ConstDiffParams — Type.

ConstDiffParams(eta, kap, kaph, nkaph, u, v)
ConstDiffParams(eta, kap, u, v)

Returns the params for constant diffusivity problem with time-varying flow.

source

FourierFlows.TracerAdvDiff.ConstDiffSteadyFlowParams — Type.

ConstDiffSteadyFlowParams(eta, kap, kaph, nkaph, u, v, g)
ConstDiffSteadyFlowParams(eta, kap, u, v, g)

Returns the params for constant diffusivity problem with time-steady flow.

source

FourierFlows.TracerAdvDiff.Vars — Method.

Vars(g)

Returns the vars for constant diffusivity problem on grid g.

source

Private types in module VerticallyFourierBoussinesq:

FourierFlows.VerticallyFourierBoussinesq.Params — Type.

Params(nu0, nnu0, nu1, nnu1, f, N, m, Ub, Vb)

Construct parameters for the Two-Fourier-mode Boussinesq problem. Suffix 0 refers to zeroth mode; 1 to first mode. f, N, m are Coriolis frequency, buoyancy frequency, and vertical wavenumber of the first mode, respectively. The optional constant background velocity (Ub,Vb) is set to zero by default. The viscosity is applied only to the first-mode horizontal velocities.

source

Private types in module VerticallyCosineBoussinesq:

FourierFlows.VerticallyCosineBoussinesq.ForcedVars — Method.

ForcedVars(g)

Returns the vars for forced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source

FourierFlows.VerticallyCosineBoussinesq.Params — Type.

Params(nu0, nnu0, nu1, nnu1, mu0, nmu0, mu1, nmu1, f, N, m; Ub=0, Vb=0)

Construct parameters for the Two-Fourier-mode Boussinesq problem. Suffix 0 refers to zeroth mode; 1 to first mode. f, N, m are Coriolis frequency, buoyancy frequency, and vertical wavenumber of the first mode, respectively. The optional constant background velocity (Ub, Vb) is set to zero by default. The viscosity is applied only to the first-mode horizontal velocities.

source

FourierFlows.VerticallyCosineBoussinesq.TracerForcedVars — Method.

TracerForcedVars(g)

Returns the vars for forced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source

FourierFlows.VerticallyCosineBoussinesq.Vars — Method.

Vars(g)

Returns the vars for unforced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source