Functions

Functions

Functions exported from FourierFlows:

Base.resize! — Method.

resize!(diag, newnum)

Resize the Diagnostic data and time arrays to length newnum.

source

FourierFlows.TimeStepper — Method.

Returns a time-stepper of type `steppernameTimeStepper'.

source

FourierFlows.dealias! — Method.

dealias!(a, g, kalias)

Dealias a on the grid g with aliased x-wavenumbers kalias.

source

FourierFlows.gridpoints — Method.

gridpoints(g)

Returns the collocation points of the grid g in 2D arrays X, Y.

source

FourierFlows.increment! — Method.

increment!(diag)
increment!(diags)

Increment the Diagnostic diag, or an array of Diagnostics diags.

source

FourierFlows.savediagnostic — Method.

savediagnostic(diag, diagname)

Save diagnostics to file, labeled by the string diagname.

source

FourierFlows.saveoutput — Method.

saveoutput(out)

Save current output fields for file in out.filename.

source

FourierFlows.saveproblem — Method.

saveproblem(prob, filename)

Save certain aspects of a problem timestepper, grid, and params. Functions that are fields in params are not saved.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, diags, nsteps)

Step forward prob for nsteps, incrementing diagnostics in the array diags along the way.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, nsteps)

Step forward prob for nsteps.

source

FourierFlows.stepforward! — Method.

stepforward!(prob)

Step forward prob for one timestep.

source

FourierFlows.update! — Method.

update!(diag)

Update diag with its current value.

source

Functions exported from KuramotoSivashinsky:

FourierFlows.KuramotoSivashinsky.InitialValueProblem — Method.

InitialValueProblem(; parameters...)

Construct an initial-value Kuramoto-Sivashinky problem that solves the equation

∂t u + ∂ₓ⁴u + ∂ₓ²u + u ∂ₓu = 0.

source

FourierFlows.KuramotoSivashinsky.set_u! — Method.

set_u!(prob, u)
set_u!(s, v, g, u)

Set the solution prob.state.sol as the transform of u and update variables.

source

FourierFlows.KuramotoSivashinsky.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

source

Functions exported from TwoDTurb:

Modules = [FourierFlows.TwoDTurb]
Private = false
Order = [:function]

Functions exported from BarotropicQG:

Modules = [FourierFlows.BarotropicQG]
Private = false
Order = [:function]

Functions exported from TracerAdvDiff:

FourierFlows.TracerAdvDiff.set_c! — Method.

set_c!(s, v, g, c)
set_c!(s, v, g, c::Function)
set_c!(prob, c)

Set the solution s.sol as the transform of c and update variables v on the grid g.

source

FourierFlows.TracerAdvDiff.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

source

Functions exported from VerticallyFourierBoussinesq:

Modules = [FourierFlows.VerticallyFourierBoussinesq]
Private = false
Order = [:function]

Functions exported from VerticallyCosineBoussinesq:

Modules = [FourierFlows.VerticallyCosineBoussinesq]
Private = false
Order = [:function]