Functions exported from FourierFlows:

Functions exported from FourierFlows:

Base.resize! — Method.

Resize the diagnostic data and time arrays.

source

FourierFlows.groupsize — Method.

Find the number of elements in a JLD2 group.

source

FourierFlows.increment! — Method.

Increment an array of diagnostics.

source

FourierFlows.increment! — Method.

Increment a diagnostic by calculating a new value.

source

FourierFlows.savediagnostic — Method.

Save diagnostics to file.

source

FourierFlows.saveoutput — Method.

Save the current output fields.

source

FourierFlows.saveproblem — Method.

Save certain aspects of a Problem. Entire problems cannot be saved in general, because functions cannot be saved (and functions may use arbitrary numbers of global variables that cannot be included in a saved object).

source

FourierFlows.saveproblem — Method.

Save attributes of the Problem associated with the given Output.

source

FourierFlows.update! — Method.

Update the current value of the diagnostic.

source

Private types in module FourierFlows:

FourierFlows.Equation — Type.

This type defines the linear implicit and explicit components of an equation. The linear implicit part of an equation is defined by an array of coefficients which multiply the solution. The explicit part of an equation is calculated by a function that may define linear and nonlinear parts.

source

FourierFlows.ZeroDGrid — Type.

Doc.

source