Functions

Functions

Functions exported from FourierFlows:

FourierFlows.TimeStepper — Function.

TimeStepper(stepper, eq, dt=nothing)

Generalized timestepper constructor. If stepper is explicit, dt is not used.

source

FourierFlows.cxtype — Method.

cxtype(T)

Returns T when T is Complex, or Complex{T} when T is Real.

source

FourierFlows.dealias! — Method.

dealias!(a, g, kalias)

Dealias a on the grid g with aliased x-wavenumbers kalias.

source

FourierFlows.devzeros — Method.

devzeros(dev, T, dims)

Returns an array like A of type T, but full of zeros.

source

FourierFlows.fltype — Method.

fltype(T)

Returns T when T<:AbstractFloat or Tf when T<:Complex{Tf}.

source

FourierFlows.gridpoints — Method.

gridpoints(g)

Returns the collocation points of the grid g in 2D arrays X, Y.

source

FourierFlows.increment! — Method.

increment!(diag)

Increment the Diagnostic diag, or an array of Diagnostics diags.

source

FourierFlows.innereltype — Method.

innereltype(x)

Recursively determine the 'innermost' type in by the collection x (which may be, for example, a collection of a collection).

source

FourierFlows.savediagnostic — Method.

savediagnostic(diag, diagname)

Save diagnostics in diag to file, labeled by diagname.

source

FourierFlows.saveoutput — Method.

saveoutput(out)

Save the fields in out.fields to out.path.

source

FourierFlows.saveproblem — Method.

saveproblem(prob, filename)

Save certain aspects of a problem.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, diags, nsteps)

Step forward prob for nsteps, incrementing diags along the way. diags may be a single Diagnostic or a Vector of Diagnostics.

source

FourierFlows.stepforward! — Method.

stepforward!(prob, nsteps)

Step forward prob for nsteps.

source

FourierFlows.stepforward! — Method.

stepforward!(prob)

Step forward prob one time step.

source

FourierFlows.superzeros — Method.

superzeros(T, A)

Returns an array like A, but full of zeros. If innereltype(A) can be promoted to T, then the innermost elements of the array will have type T.

source

FourierFlows.update! — Method.

update!(diag)

Update diag.

source