Private types

Private types

Private types in module FourierFlows:

FourierFlows.Equation — Type.

This type defines the linear implicit and explicit components of an equation. The linear implicit part of an equation is defined by an array of coefficients which multiply the solution. The explicit part of an equation is calculated by a function that may define linear and nonlinear parts.

source

FourierFlows.Problem — Type.

Problem(g, v, p, eq, ts)

Initialize a FourierFlows problem on grid g, with variables v, parameters p, equation eq, and timestepper ts.

source

FourierFlows.ZeroDGrid — Type.

ZeroDGrid()

Constructs a placeholder grid object for "0D" problems (in other words, systems of ODEs).

source

Private types in module KuramotoSivashinsky:

FourierFlows.KuramotoSivashinsky.Vars — Method.

Returns the Vars object for Kuramoto-Sivashinsky.

source

Private types in module TwoDTurb:

Modules = [FourierFlows.TwoDTurb]
Public = false
Order = [:type]

Private types in module BarotropicQG:

Modules = [FourierFlows.BarotropicQG]
Public = false
Order = [:type]

Private types in module TracerAdvDiff:

FourierFlows.TracerAdvDiff.ConstDiffParams — Type.

ConstDiffParams(eta, kap, kaph, nkaph, u, v)
ConstDiffParams(eta, kap, u, v)

Returns the params for constant diffusivity problem with time-varying flow.

source

FourierFlows.TracerAdvDiff.ConstDiffSteadyFlowParams — Type.

ConstDiffSteadyFlowParams(eta, kap, kaph, nkaph, u, v, g)
ConstDiffSteadyFlowParams(eta, kap, u, v, g)

Returns the params for constant diffusivity problem with time-steady flow.

source

FourierFlows.TracerAdvDiff.Vars — Method.

Vars(g)

Returns the vars for constant diffusivity problem on grid g.

source

Private types in module VerticallyFourierBoussinesq:

Modules = [FourierFlows.VerticallyFourierBoussinesq]
Public = false
Order = [:type]

Private types in module VerticallyCosineBoussinesq:

Modules = [FourierFlows.VerticallyCosineBoussinesq]
Public = false
Order = [:type]